Refactor all modules
